Created separate class SymmetryFinder
